+++ ezgradle-plugin/src/main/java/me/thens/ezgradle/model/ProjectProperties.kt
@@ -9,33 +9,12 @@ class ProjectProperties(val project: Project, val prefix: String = "") {
         return ProjectProperties(project, "$prefix$nextPrefix")
     }
 
-    fun getString(name: String): String? {
+    operator fun get(name: String): String? {
         val key = "$prefix$name"
-        return project.findProject(key)?.toString()
-            ?: System.getenv(key.toEnvName())
-    }
-
-    fun getBoolean(name: String): Boolean? {
-        return getString(name)?.toBoolean()
-    }
-
-    fun getInt(name: String): Int? {
-        return getString(name)?.toIntOrNull()
-    }
-
-    fun getLong(name: String): Long? {
-        return getString(name)?.toLongOrNull()
-    }
-
-    fun getFloat(name: String): Float? {
-        return getString(name)?.toFloatOrNull()
-    }
-
-    fun getDouble(name: String): Double? {
-        return getString(name)?.toDoubleOrNull()
+        return System.getenv(key.toEnvName()) ?: project.findProject(key)?.toString()
     }
 
     fun <E : Enum<E>> getEnum(name: String, defaultValue: E): E {
-        return getString(name)?.toEnum(defaultValue.javaClass) ?: defaultValue
+        return get(name)?.toEnum(defaultValue.javaClass) ?: defaultValue
     }
 }